+++ projects/angular-sdk/src/apis/hopper-cloud-airline/v1/model/cfarContractCustomer.ts
@@ -12,4 +12,5 @@
 
 export interface CfarContractCustomer { 
     id: string;
+    premium: string;
 }