+++ projects/angular-sdk/CHANGELOG.md
@@ -1,3 +1,12 @@
+<a name="2.10.0"></a>
+
+# 2.10.0 (2022-09-07)
+
+### Minor changes
+- Offer Banner Large:
+    - Improved the display of the error message when there are no offers
+    - Transmission of the premium price when selecting an offer
+
 <a name="2.9.0"></a>
 
 # 2.9.0 (2022-08-30)